requirements.txt
python-dotenv
playwright
requests
openai
fastapi
uvicorn[standard]
jinja2
aiofiles
python-socks
